Switch scripts to use Tap.fetch
This removes the redundant tap_name logic from the script (which already
exists in homebrew itself). Even better, this will leverage the tap
cache! (And it still preserves the CoreTap singleton preference.)

Also, update task args to use the hash-index form consistently
throughout the rakefile, instead of the (potentially brittle) method
form (which is susceptible to collisions with existing methods).