Ensure linux_formula_and_analytics task runs
The formulae and analytics tasks are frequently run for macOS prior to
the Linux variant running. And rake considers the tasks already
completed, so they are then skipped for linux.

This re-enables the formulae and analytics tasks so that they'll run
again for Linux. Technically this means it is now possible for the linux
tasks to be run multiple times (unintentionally), but I believe that's a
less-bad outcome than not running when expected.