Add generated-data directories to rake's CLOBBER list
As all the formula, cask, and analytics data files are generated by
various rake tasks, it is also desirable to be able to start "clean".

The scripts that generate these files already safely remove/recreate the
directory structures, but adding them explicitly to rake's CLOBBER list
enables users to get to a "clean" working directory (for inspection,
or other purposes) before regenerating them.