Allow use of rake without a gem bundle
This allows the data generation tasks to be run outside the context of a
Bundler gem bundle. This is necessary for the core and cask taps to
trigger data re-generation from their Action workflows.

Hopefully this is a short-term solution, as it's preferable for rake to
be able to assume (and leverage) access to its bundled gems.

see https://github.com/Homebrew/formulae.brew.sh/pull/271#issuecomment-615897164