Start using internal formula json
This is the slimmer JSON that is missing certain keys that won't be missed.
The goal is that this smaller file contains the same amount of information
as the previous file but means less information is flying around the internet
and it will help speed up JSON parsing times as well since there's less to
parse.